+++ src/index.tsx
@@ -1,42 +1,11 @@
 import { createRoot } from 'react-dom/client';
-import { StrictMode, CSSProperties, useState } from 'react';
-import clsx from 'clsx';
-
-import { Article } from './components/article/Article';
-import { ArticleParamsForm } from './components/article-params-form/ArticleParamsForm';
-import {
-	ArticleStateType,
-	defaultArticleState,
-} from './constants/articleProps';
+import { StrictMode } from 'react';
 
 import './styles/index.scss';
-import styles from './styles/index.module.scss';
+import { App } from './components/app/App';
 
 const domNode = document.getElementById('root') as HTMLDivElement;
 const root = createRoot(domNode);
-
-const App = () => {
-	const [params, setParams] = useState(defaultArticleState);
-	function handleChange(newParams: ArticleStateType) {
-		setParams(newParams);
-	}
-	return (
-		<div
-			className={clsx(styles.main)}
-			style={
-				{
-					'--font-family': params.fontFamilyOption.value,
-					'--font-size': params.fontSizeOption.value,
-					'--font-color': params.fontColor.value,
-					'--container-width': params.contentWidth.value,
-					'--bg-color': params.backgroundColor.value,
-				} as CSSProperties
-			}>
-			<ArticleParamsForm onChange={handleChange} />
-			<Article />
-		</div>
-	);
-};
 
 root.render(
 	<StrictMode>
